Fix library management CI regressions
